requirements.txt
aiogram==3.*
aiohttp>=3.9,<4.0
pydantic>=2.1,<3
pydantic-settings>=2.1,<3
sqlalchemy>=2.0,<3.0
alembic>=1.13,<2.0
python-dotenv>=1.0,<2.0
